Moved slash command management in a new file
Continuing to clean up the xml_to_js file
